do not run disable-p2p if nm not installed

+++ utils/hooks/disable-p2pdev.sh
@@ -1,5 +1,9 @@
+if [ -f $ROOTFS/etc/NetworkManager/NetworkManager.conf ]; then
+
 cat <<EOF | sudo tee -a $ROOTFS/etc/NetworkManager/NetworkManager.conf
 
 [keyfile]
 unmanaged-devices=interface-name:p2p0
 EOF
+
+fi
